requirements.txt
Django==4.2.7
Pillow==10.0.1
tensorflow==2.19.0
numpy==2.1.3
opencv-python==4.8.1.78
requests==2.31.0
python-decouple==3.8
django-cors-headers==4.3.1 